Moving compute function to pvnet_uk_regional.py and Rename process_and_combine to merge_and_fill.py
